+++ packages/saved-views-react/CHANGELOG.md
@@ -45,6 +45,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Fixes
 
+* Fix `captureSavedViewData` failing with blank iModel connections
 * Replace usage of internal iTwin.js API with public alternative
 
 ## [0.6.0](https://github.com/iTwin/saved-views/tree/react-v0.6.0/packages/saved-views-react/) - 2024-07-22